Bug fix in KBI for simulations with very large number of particles (changed 32-bit integer to 64-bit integer of KBISum).
git-svn-id: https://scm.projects.hlrs.de/authscm/sleyh/svn/ms2/trunk@580 a2be33b2-7eaf-4dce-bf61-b02fde6c56c2

+++ src/ms2_interaction.F90
@@ -61,7 +61,7 @@ module ms2_interaction
     real(RK), pointer, contiguous :: MayerFFunction1(:), IntFFunction1(:)
     real(RK), pointer, contiguous :: MayerFFunction2(:), IntFFunction2(:)
     ! Array for the shells of RDF inside of KBI
-    integer, pointer, contiguous          :: KBISum(:)
+    integer(KIND=8), pointer, contiguous          :: KBISum(:)
 
     ! Virial
     real(RK), pointer, contiguous :: Virial(:, :), Virial1(:), VirialNew(:, :)
